Clearly fail when parsing fails
Trouble I ran into trying to parse the imgui.h in my project.

If we fail to parse, we should fail with a message instead of from
passing nil to gsub:
    generator.lua:324: bad argument #2 to 'gsub' (string/function/table expected)

imconfig is likely to require other headers that aren't in our include
path. Commenting out the include or passing
/DIMGUI_DISABLE_INCLUDE_IMCONFIG_H to generator.lua is an option.